shopglobal_id_search/requirements.txt
python-dotenv==1.0.1
requests==2.31.0
black==23.11.0
flake8==6.1.0
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
gunicorn==21.2.0
python-multipart==0.0.9
google-cloud-bigquery==3.13.0
google-auth==2.27.0 